use prisma-client-rust main branch (#2471)
* use primsa-client-rust main branch

* remove as i64

+++ core/src/object/media/old_media_processor/job.rs
@@ -602,7 +602,7 @@ async fn get_files_for_labeling(
 				""
 			}
 		),
-		PrismaValue::Int(parent_iso_file_path.location_id() as i64),
+		PrismaValue::Int(parent_iso_file_path.location_id()),
 		PrismaValue::String(format!(
 			"{}%",
 			parent_iso_file_path
@@ -640,7 +640,7 @@ async fn get_all_children_files_by_extensions(
 				.collect::<Vec<_>>()
 				.join(",")
 		),
-		PrismaValue::Int(parent_iso_file_path.location_id() as i64),
+		PrismaValue::Int(parent_iso_file_path.location_id()),
 		PrismaValue::String(format!(
 			"{}%",
 			parent_iso_file_path
